+++ server/helpers/dbImport/mappings.ts
@@ -3,8 +3,7 @@ import { Import } from "./import";
 export default class Mappings extends Import {
   display = "Mappings";
   
-  service = 0;
-  inputTable = () => `current_mappings_${this.service}`;
+  inputTable = () => `current_mappings_${this.service?.id}`;
   
   initialKey = [-1, -1];
   outputTable = "mappings";
@@ -21,25 +20,4 @@ export default class Mappings extends Import {
     ORDER BY mappings.hash_id, mappings.tag_id
     LIMIT ?
   `;
-  
-  async startEach(services: number[]) {
-    const sizes = services.map(service => {
-      this.resetTotal();
-      this.service = service;
-      return { service, total: this.total() };
-    });
-    
-    sizes.sort((a, b) => b.total - a.total);
-    
-    for(const { service, total } of sizes) {
-      if(total === 0) continue;
-      
-      this.resetTotal(total);
-      this.service = service;
-      
-      await this.start();
-      
-      this.useTemp = true;
-    }
-  }
 }