+++ server/helpers/dbImport/posts.ts
@@ -1,12 +1,12 @@
 import { PoolClient } from "pg";
 import { Database } from "better-sqlite3";
+import { ServiceID } from "../consts";
 import { Import } from "./import";
 import { Service } from "./index";
 
 export default class Posts extends Import {
   display = "Posts";
   
-  service?: Service;
   inputTable = () => `current_files_${this.service?.id || 0}`;
   
   batchSizeMul = 1 / 2;
@@ -29,42 +29,21 @@ export default class Posts extends Import {
       COALESCE(local_ratings.rating, '') || ',' ||
       COALESCE(files_info.mime, '') || ',' ||
       (file_inbox.hash_id IS NOT NULL) || ',' ||
-      ${!!this.service?.trash} || ',' ||
+      ${this.service?.type === ServiceID.LOCAL_FILE_TRASH_DOMAIN} || ',' ||
       datetime(current_files.timestamp_ms / 1000, 'unixepoch', 'utc') || '\n'
     FROM ${this.inputTable()} current_files
       INNER JOIN hashes ON hashes.hash_id = current_files.hash_id
       LEFT JOIN files_info ON files_info.hash_id = current_files.hash_id
       LEFT JOIN local_hashes ON local_hashes.hash_id = current_files.hash_id
       LEFT JOIN blurhashes ON blurhashes.hash_id = current_files.hash_id
-      LEFT JOIN local_ratings ON local_ratings.service_id = ${this.ratingService} AND local_ratings.hash_id = current_files.hash_id
+      LEFT JOIN local_ratings ON local_ratings.service_id = ${this.ratingService?.id || null} AND local_ratings.hash_id = current_files.hash_id
       LEFT JOIN file_inbox ON file_inbox.hash_id = current_files.hash_id
     WHERE current_files.hash_id > ?
     ORDER BY current_files.hash_id ASC
     LIMIT ?
   `;
   
-  constructor(hydrus: Database, postgres: PoolClient, public ratingService: number | null) {
+  constructor(hydrus: Database, postgres: PoolClient, public ratingService: Service | null) {
     super(hydrus, postgres);
-  }
-  
-  async startEach(services: Service[]) {
-    const sizes = services.map(service => {
-      this.resetTotal();
-      this.service = service;
-      return { service, total: this.total() };
-    });
-    
-    sizes.sort((a, b) => b.total - a.total);
-    
-    for(const { service, total } of sizes) {
-      if(total === 0) continue;
-      
-      this.resetTotal(total);
-      this.service = service;
-      
-      await this.start();
-      
-      this.useTemp = true;
-    }
   }
 }